change some frequently used data types from double to float

+++ GlobalQuakeCore/src/main/java/globalquake/core/station/GlobalStationManager.java
@@ -66,7 +66,7 @@ public class GlobalStationManager {
                         double ang = GeoUtils.calculateAngle(stat.getLatitude(), stat.getLongitude(), stat2.getLatitude(),
                                 stat2.getLongitude());
                         int ray = (int) ((ang / 360.0) * (RAYS - 1.0));
-                        rays.get(ray).add(new StationDistanceInfo(stat2.getId(), dist, ang));
+                        rays.get(ray).add(new StationDistanceInfo(stat2.getId(), (float) dist, (float) ang));
                         int ray2 = ray + 1;
                         if (ray2 == RAYS) {
                             ray2 = 0;
@@ -75,8 +75,8 @@ public class GlobalStationManager {
                         if (ray3 == -1) {
                             ray3 = RAYS - 1;
                         }
-                        rays.get(ray2).add(new StationDistanceInfo(stat2.getId(), dist, ang));
-                        rays.get(ray3).add(new StationDistanceInfo(stat2.getId(), dist, ang));
+                        rays.get(ray2).add(new StationDistanceInfo(stat2.getId(), (float) dist, (float) ang));
+                        rays.get(ray3).add(new StationDistanceInfo(stat2.getId(), (float) dist, (float) ang));
                         num++;
                     }
                 }
@@ -131,7 +131,7 @@ public class GlobalStationManager {
     }
 
 
-    record StationDistanceInfo(int id, double dist, double ang) {
+    record StationDistanceInfo(int id, float dist, float ang) {
 
     }
